clippy: unbreak `zcore-drivers` and the libos build on master
Two lints landed under `#![deny(warnings)]` and have been failing CI for
everyone since: an `as u32` on a value that is already `u32` in the e1000e
NVM bank detect, and an early-return block clippy wants as `?` in the
syncobj sync_file poll path. Both are the compiler's own suggestion applied
as is, with no behaviour change; `cargo clippy --package zcore --features
"linux libos"` and `cargo clippy -p zcore-drivers --lib --features graphic`
are clean again.

+++ drivers/src/net/e1000e.rs
@@ -1468,9 +1468,10 @@ impl E1000eHw {
     /// Bank detect via signature word 0x13 (Linux `e1000_valid_nvm_bank_detect_ich8lan` SPT).
     unsafe fn ich_flash_active_bank_words(&self) -> u32 {
         let strap = mmio_read(self.base, E1000E_STRAP);
-        let nvm_size = ((((strap >> 1) & 0x1F) + 1) * 4096) as u32;
+        let nvm_size = (((strap >> 1) & 0x1F) + 1) * 4096;
         let bank_words = (nvm_size / 2) / 2; // words per bank
-                                             // Signature at word offset 0x13 of each bank; valid if (sig & 0xC0) == 0x80.
+
+        // Signature at word offset 0x13 of each bank; valid if (sig & 0xC0) == 0x80.
         if let Some(d) = self.flash_read_dword_spt(0x13 << 1) {
             let sig = ((d >> 8) & 0xFF) as u16;
             if sig & 0xC0 == 0x80 {
